% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_eavs.R
\name{create_eavs_years}
\alias{create_eavs_years}
\title{create_eavs_years()}
\usage{
create_eavs_years(years = seq(2008, 2020, 2), eavs.dir = "data/")
}
\arguments{
\item{years}{Array of valid years (even years from 2008-2020).}

\item{eavs.dir}{Defaults to data. Should be the extract.dir argument from download_eavs_year().}
}
\value{
a data frame.
}
\description{
Create a multi-year EAVS file.
}
