% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_eavs.R
\name{download_eavs}
\alias{download_eavs}
\title{download_eavs()}
\usage{
download_eavs(years = "", extract.dir = paste0(getwd(), "/data"))
}
\arguments{
\item{years}{Array of valid years (even years from 2008-2020) or "all" or blank for all years.}

\item{extract.dir}{Defaults to data in current directory.}
}
\value{
a data frame.
}
\description{
Download EAVS data directly from the Election Assistance Commission. Note: we will be mirroring this data elsewhere for stability.
}
