% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eavs_var.R
\name{eavs_var}
\alias{eavs_var}
\title{download_eavs()}
\usage{
eavs_var(var = "", question = "", topic = "", print.out = T)
}
\arguments{
\item{var}{Searches variable names for matches. (example 'qa1a')}

\item{question}{Searches question text. (example "registrations")}

\item{topic}{Searches topics. List of topics can be seen with list_topics() function.}

\item{print.out}{logical. FALSE returns data.frame of variables.}
}
\value{
a data frame or printed output.
}
\description{
Used to find and define EAVS variables in the clean EAVS file, or the regular EAVS
}
