% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_eavs.R
\name{fix_types}
\alias{fix_types}
\title{fix_types()}
\usage{
fix_types(x)
}
\arguments{
\item{x}{Data.frame. Output of rename_eavs() function.}
}
\value{
a data frame, with dimensions depending on the year.
}
\description{
Year agnostic data types for binding rows of eavs
}
