% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_eavs.R
\name{read_eavs}
\alias{read_eavs}
\title{read_eavs()}
\usage{
read_eavs(year, data.dir = "data")
}
\arguments{
\item{year}{Which year is being read; defaults to 4-digit year in file name}

\item{data.dir}{directory that defaults to data/ in current directory}
}
\value{
a data frame, with dimensions depending on the year.
}
\description{
Read one year of data from the Election Administration and Voting Survey
}
