% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_eavs.R
\name{rename_eavs}
\alias{rename_eavs}
\title{rename_eavs()}
\usage{
rename_eavs(eavs)
}
\arguments{
\item{eavs}{Data.frame. Output of read_eavs() function.}
}
\value{
a data frame, with dimensions depending on the year.
}
\description{
Year agnostic names for the EAVS
}
